:truck: [#1183] -- move utility to registration app from stuf-zds plugin
